feat(product): MenuPrice 일급 객체 생성
