+++ src/main/java/kitchenpos/domain/Product.java
@@ -1,18 +1,14 @@
 package kitchenpos.domain;
 
 import java.math.BigDecimal;
-import java.util.Objects;
 
 public class Product {
-
-    private static final int MIN_PRICE = 0;
 
     private Long id;
     private String name;
     private BigDecimal price;
 
     public Product(Long id, String name, BigDecimal price) {
-        check(price);
         this.id = id;
         this.name = name;
         this.price = price;
@@ -24,12 +20,6 @@ public class Product {
 
     public Product(String name, int price) {
         this(null, name, new BigDecimal(price));
-    }
-
-    private void check(BigDecimal price) {
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < MIN_PRICE) {
-            throw new IllegalArgumentException();
-        }
     }
 
     public Long getId() {
